requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
geopy==2.4.1
shapely==2.0.5
h3==3.7.7
scikit-learn==1.5.1
jinja2==3.1.4
# Optional for advanced ML (comment if not needed)
# xgboost==2.1.1
# lightgbm==4.5.0
# Optional for advanced vision (comment if not needed)
# torch
# open-clip-torch
streamlit==1.38.0
